Extract class CircleAnimation
